Solution of most recent blocking issues with brat and brat-scheduler starting development
Generic:
- Warning elimination in legacy code
- Update of VM windows 64 bit development

Work on analysis application:
- Extended include of display directory in brathlGuiUtilities to all nixes.
- Qwt integration in test application.  Added 2D plot files to qtc settings
- Analysis and solution of the problem with QGIS layers registry.

Work on scheduler:
- Complete CBratTask XML serialization.
-  BratScheduler analysis with code rewriting - replaced wx XML document inheritance, preparing migration of orchestrator (config) class

+++ libbrathl/Parameter.cpp
@@ -391,7 +391,7 @@ void CParameter::GetValue(std::string& value, int32_t pos /* = 0*/, const std::s
 }
 
 //----------------------------------------
-void CParameter::GetValue(char* value, int32_t bufferSize, int32_t pos /* = 0*/, const char *DefValue /*=""*/)
+void CParameter::GetValue(char* value, size_t bufferSize, int32_t pos /* = 0*/, const char *DefValue /*=""*/)
 {
   memset(value, '\0', bufferSize);
