+++ libbrathl/ProductEnvisat.cpp
@@ -659,7 +659,7 @@ void CProductEnvisat::ComputeHighResolutionFields(CDataSet* dataSet)
 
   size_t count = dataSet->size();
 
-  for (int32_t index = 0 ; index < count ; index++)
+  for (size_t index = 0 ; index < count ; index++)
   {
     dataSet->SetCurrentRecordSet(index);
 
@@ -1055,7 +1055,7 @@ void CProductEnvisat::ComputeHighResolutionFields(CDataSet* dataSet, double delt
 
   size_t count = dataSet->size();
 
-  for (int32_t index = 0 ; index < count ; index++)
+  for (size_t index = 0 ; index < count ; index++)
   {
     dataSet->SetCurrentRecordSet(index);
