+++ libbrathl/ProductErsWAP.cpp
@@ -413,7 +413,7 @@ void CProductErsWAP::ComputeHighResolutionFields(CDataSet* dataSet)
 
   size_t count = dataSet->size();
 
-  for (int32_t index = 0 ; index < count ; index++)
+  for (size_t index = 0 ; index < count ; index++)
   {
     dataSet->SetCurrentRecordSet(index);
