+++ libbrathl/ProductJason.cpp
@@ -505,7 +505,7 @@ void CProductJason::ComputeHighResolutionFields(CDataSet* dataSet, double deltaL
 
   size_t count = dataSet->size();
 
-  for (int32_t index = 0 ; index < count ; index++)
+  for (size_t index = 0 ; index < count ; index++)
   {
     dataSet->SetCurrentRecordSet(index);
