+++ libbrathl/ProductTopex.cpp
@@ -552,7 +552,7 @@ void CProductTopex::ComputeHighResolutionFields(CDataSet* dataSet, double deltaL
 
   size_t count = dataSet->size();
 
-  for (int32_t index = 0 ; index < count ; index++)
+  for (size_t index = 0 ; index < count ; index++)
   {
     dataSet->SetCurrentRecordSet(index);
