+++ libbrathl/ProductTopexSDR.cpp
@@ -403,7 +403,7 @@ void CProductTopexSDR::ComputeHighResolutionFields(CDataSet* dataSet, double del
 
   size_t count = dataSet->size();
 
-  for (int32_t index = 0 ; index < count ; index++)
+  for (size_t index = 0 ; index < count ; index++)
   {
     dataSet->SetCurrentRecordSet(index);
